Refactored delete documents into its own file
Ran tensor search unit tests - passed (besides a randomly failing one)
