+++ examples/StableDiffusion/hot-dog-100k.md
@@ -92,6 +92,8 @@ results = client.index("hot-dogs-100k").search("a black image")
 Now we have a black image, we can search using that and find all the other duplicate black images and remove them from the dataset.
 
 ```python
+import marqo.tensor_search.delete_docs
+
 query = 'a black image'
 
 results = client.index(index_name).search(query)
@@ -102,7 +104,7 @@ results = client.index(index_name).search(results['hits'][0]['image_docker'], li
 # we check the results - scores of very close to 1 are duplicated (this value can change depending on the task)
 documents_delete = [r['_id'] for r in results['hits'] if r['_score'] > 0.99999]
 
-client.index(index_name).delete_documents(documents_delete)
+marqo.tensor_search.delete_docs.delete_documents(documents_delete)
 ```
 
 Now our dataset should be free from images that do not contain hot dogs.
@@ -179,22 +181,24 @@ To animate the images based on their sorted vectors we can take an image as a st
 
 ```python
 # pick one to start
-results = client.index("hot-dogs-100k").search('a photo of a smiling face', 
-                      searchable_attributes=['image_docker'], filter_string="a_face:[0.58 TO 0.99]")
+import marqo.tensor_search.delete_docs
+
+results = client.index("hot-dogs-100k").search('a photo of a smiling face',
+                                               searchable_attributes=['image_docker'],
+                                               filter_string="a_face:[0.58 TO 0.99]")
 # find the document that matches closest with the query
-index = [ind for ind,doc in enumerate(documents) if doc['_id'] == results['hits'][0]['_id'] ][0]
+index = [ind for ind, doc in enumerate(documents) if doc['_id'] == results['hits'][0]['_id']][0]
 current_document = documents[index]
 # create a list to store the "sorted" documents
 ordered_documents = [current_document['_id']]
 
 for i in range(len(documents)):
-
     # remove current document
-    client.index(index_name).delete_documents([current_document['_id']])
+    marqo.tensor_search.delete_docs.delete_documents([current_document['_id']])
 
     # now search with it to get next best
     results = client.index(index_name).search(current_document['image_docker'], filter_string="a_face:[0.58 TO 0.99]",
-                            searchabel_attributes=['image_docker'], device='cuda')
+                                              searchabel_attributes=['image_docker'], device='cuda')
 
     next_document = results['hits'][0]
