+++ src/marqo/tensor_search/tensor_search.py
@@ -32,14 +32,13 @@ Notes on search behaviour with caching and searchable attributes:
 """
 import copy
 import json
-import datetime
 from collections import defaultdict
 from timeit import default_timer as timer
 import functools
 import pprint
 import typing
 import uuid
-from typing import List, Optional, Union, Iterable, Sequence, Dict, Any, Tuple, Set
+from typing import List, Optional, Union, Iterable, Sequence, Dict, Any, Tuple
 import numpy as np
 from PIL import Image
 import marqo.config as config
@@ -957,41 +956,6 @@ def _get_documents_for_upsert(
 
     # Returns a list of combined docs
     return res
-
-
-def delete_documents(config: Config, index_name: str, doc_ids: List[str], auto_refresh):
-    """Deletes documents """
-    if not doc_ids:
-        raise errors.InvalidDocumentIdError("doc_ids can't be empty!")
-
-    for _id in doc_ids:
-        validation.validate_id(_id)
-
-    # TODO: change to timer()
-    t0 = datetime.datetime.utcnow()
-    delete_res_backend = HttpRequests(config=config).post(
-        path=f"{index_name}/_delete_by_query", body={
-            "query": {
-                "terms": {
-                    "_id": doc_ids
-                }
-            }
-        }
-    )
-    if auto_refresh:
-        refresh_response = HttpRequests(config).post(path=F"{index_name}/_refresh")
-    t1 = datetime.datetime.utcnow()
-    delete_res = {
-        "index_name": index_name, "status": "succeeded",
-        "type": "documentDeletion", "details": {
-            "receivedDocumentIds": len(doc_ids),
-            "deletedDocuments": delete_res_backend["deleted"],
-        },
-        "duration": utils.create_duration_string(t1 - t0),
-        "startedAt": utils.format_timestamp(t0),
-        "finishedAt": utils.format_timestamp(t1),
-    }
-    return delete_res
 
 
 def refresh_index(config: Config, index_name: str):
